Rename and simplify procedure version for openhtf script
Remove url
